Rename NamedTree to AttrTree and move it to utils
