+++ my/utils/__init__.py
@@ -1,0 +1,1 @@
+from .tree import AttrTree, AttrTreeConfig
